refactor: pydanticify the tls responses
These are going to go out over routers and thus need to be actual
models. They're not used internally so no reason to keep them dataclasses.
